feat: change request status from the UI
Modifications in frontend and backend files, allowing users to update the request status through the user interface. Changes include enabling a clickable button to indicate whether a request is resolved or needs to be reopened. Changes made to the Request service and Controller in the backend allow a PUT request to change the status of a request. Added RequestStatus enum to the frontend requests model. Changes to the RequestView component show the current request status on the page. Swagger docs updated to reflect new endpoint.

The purpose of these changes is to enable users to easily change the status of a request directly within the UI, potentially improving workflow and productivity.