style(front): improve visual feedback and navigation experience
Visual feedback in the RequestView component has been enhanced for better user experience. The request status tag color intensity was changed to provide clearer distinction, and the format of the primary information display was adjusted for better readability. The isLoading state was added to the UserRequestActions component to provide feedback while the update request status is processing. Default navigation paths were added in userRoutes and adminRoutes to increase user-friendliness. An optional titleElement property was introduced to the ContentLayout component providing more flexibility in defining the title of the elements.