feat: add support for archive feature
The code updates enable the addition of archive routing and its support in the admin panel. This includes creation and retrieval of archive entries, setting up API for interacting with the archive, and components for display in the front-end. An ArchiveController is established to delegate the tasks. The update also features revision of file names for clarity.